Moved taylor to toplevel
