--- docs/api_docs/solvers/taylor/affine.md
@@ -1,1 +1,0 @@
-::: probdiffeq.solvers.taylor.affine
