--- docs/api_docs/solvers/taylor/autodiff.md
@@ -1,1 +1,0 @@
-::: probdiffeq.solvers.taylor.autodiff
