--- docs/api_docs/solvers/taylor/estim.md
@@ -1,1 +1,0 @@
-::: probdiffeq.solvers.taylor.estim
