+++ docs/benchmarks/taylor_node/run_taylor_node.py
@@ -16,7 +16,7 @@ from diffeqzoo import backend
 from jax import config
 
 from probdiffeq.impl import impl
-from probdiffeq.solvers.taylor import autodiff
+from probdiffeq.taylor import autodiff
 from probdiffeq.util.doc_util import info
