+++ docs/examples_parameter_estimation/physics_enhanced_regression_2.md
@@ -104,7 +104,7 @@ from probdiffeq.util.doc_util import notebook
 from probdiffeq.solvers import uncalibrated, solution
 from probdiffeq.solvers.strategies.components import corrections, priors
 from probdiffeq.solvers.strategies import filters
-from probdiffeq.solvers.taylor import autodiff
+from probdiffeq.taylor import autodiff
 ```
 
 ```python
