+++ docs/examples_solver_config/dynamic_output_scales.md
@@ -31,7 +31,7 @@ from jax.config import config
 
 from probdiffeq import ivpsolve
 from probdiffeq.impl import impl
-from probdiffeq.solvers.taylor import affine
+from probdiffeq.taylor import affine
 from probdiffeq.util.doc_util import notebook
 from probdiffeq.solvers import calibrated
 from probdiffeq.solvers.strategies import filters
