+++ docs/examples_solver_config/posterior_uncertainties.md
@@ -26,7 +26,7 @@ from probdiffeq import ivpsolve, adaptive, timestep
 from probdiffeq.impl import impl
 from probdiffeq.util.doc_util import notebook
 from probdiffeq.solvers import calibrated, solution, markov
-from probdiffeq.solvers.taylor import autodiff
+from probdiffeq.taylor import autodiff
 from probdiffeq.solvers.strategies import filters, smoothers, fixedpoint
 from probdiffeq.solvers.strategies.components import corrections, priors
 ```
