+++ docs/examples_solver_config/second_order_problems.md
@@ -26,7 +26,7 @@ from probdiffeq.impl import impl
 from probdiffeq import adaptive, ivpsolve
 from probdiffeq.util.doc_util import notebook
 from probdiffeq.solvers import calibrated
-from probdiffeq.solvers.taylor import autodiff
+from probdiffeq.taylor import autodiff
 from probdiffeq.solvers.strategies.components import corrections, priors
 from probdiffeq.solvers.strategies import filters
 ```
