+++ docs/examples_solver_config/smoothing.md
@@ -30,7 +30,7 @@ from probdiffeq import ivpsolve, adaptive
 from probdiffeq.impl import impl
 from probdiffeq.util.doc_util import notebook
 from probdiffeq.solvers import calibrated, solution
-from probdiffeq.solvers.taylor import autodiff
+from probdiffeq.taylor import autodiff
 from probdiffeq.solvers.strategies.components import corrections, priors
 from probdiffeq.solvers.strategies import filters, smoothers, fixedpoint
 ```
