+++ docs/getting_started/easy_example.md
@@ -12,7 +12,7 @@ from probdiffeq.impl import impl
 from probdiffeq.solvers import uncalibrated
 from probdiffeq.solvers.strategies import smoothers
 from probdiffeq.solvers.strategies.components import corrections, priors
-from probdiffeq.solvers.taylor import autodiff
+from probdiffeq.taylor import autodiff
 
 config.update("jax_platform_name", "cpu")
 ```
