+++ probdiffeq/taylor/affine.py
