+++ probdiffeq/taylor/estim.py
