+++ tests/test_solvers/test_calibrated/test_mle_calibration_vs_calibrationfree.py
@@ -12,7 +12,7 @@ from probdiffeq.impl import impl
 from probdiffeq.solvers import calibrated, solution, uncalibrated
 from probdiffeq.solvers.strategies import filters, fixedpoint
 from probdiffeq.solvers.strategies.components import corrections, priors
-from probdiffeq.solvers.taylor import autodiff
+from probdiffeq.taylor import autodiff
 from tests.setup import setup
