+++ tests/test_solvers/test_misc/test_warnings_for_wrong_strategies.py
@@ -7,7 +7,7 @@ from probdiffeq.impl import impl
 from probdiffeq.solvers import uncalibrated
 from probdiffeq.solvers.strategies import fixedpoint, smoothers
 from probdiffeq.solvers.strategies.components import corrections, priors
-from probdiffeq.solvers.taylor import autodiff
+from probdiffeq.taylor import autodiff
 from tests.setup import setup
