+++ tests/test_solvers/test_strategies/test_smoother_vs_fixedpoint_equivalence.py
@@ -11,7 +11,7 @@ from probdiffeq.impl import impl
 from probdiffeq.solvers import solution, uncalibrated
 from probdiffeq.solvers.strategies import fixedpoint, smoothers
 from probdiffeq.solvers.strategies.components import corrections, priors
-from probdiffeq.solvers.taylor import autodiff
+from probdiffeq.taylor import autodiff
 from tests.setup import setup
