+++ tests/test_taylor/data/three_body_first_solution.npy
