+++ tests/test_taylor/data/van_der_pol_second_solution.npy
