+++ tests/test_taylor/test_inexact_first_order.py
@@ -4,7 +4,7 @@ import jax.numpy as jnp
 
 from probdiffeq.backend import testing
 from probdiffeq.impl import impl
-from probdiffeq.solvers.taylor import autodiff, estim
+from probdiffeq.taylor import autodiff, estim
 
 
 @testing.case()
